NewVoiceChannelBot/requirements.txt
discord-py-interactions==4.3.1
dataclasses_json==0.5.7
git+https://github.com/interactions-py/voice